Break task dealer into its own file
